==> notebooks/1-member_processing.ipynb <==
# 1. Member Data Import and Information Processing

This notebook will load the member data and do pre-processing on it.

## Part a: Data import
First, load the member information data from a CSV file.

import pandas as pd

# Step 1: Data Loading
# Load the members' information from the CSV file
members_df = pd.read_csv('../data/raw/Members.csv')


# display data set
members_df

## Part b: Members' gender information processing
We process the `Sex` column:
- Fill in missing gender.
- Perform One-Hot encoding.

# Members' gender Information Processing

# Fill missing values in the 'Sex' column with 'Unknown'
members_df['Sex'].fillna('Unknown', inplace=True)

# One-Hot encode the 'Sex' column into 'Male', 'Female', and 'Unknown'
members_df['Male'] = (members_df['Sex'] == 'M').astype(int)
members_df['Female'] = (members_df['Sex'] == 'F').astype(int)
members_df['Unknown'] = (members_df['Sex'] == 'Unknown').astype(int)

# Drop the original 'Sex' column
members_df.drop(columns=['Sex'], inplace=True)

## Part c: Members' age information processing for the first claim
We process the `AgeAtFirstClaim` column:
- map the age range to the average value
- Remove missing values

""" # Remove rows where the 'AgeAtFirstClaim' column is missing
members_df = members_df[members_df['AgeAtFirstClaim'].notna()] """

# Map age ranges to their respective average values
age_mapping = {
    '0-9': 5,
    '10-19': 15,
    '20-29': 25,
    '30-39': 35,
    '40-49': 45,
    '50-59': 55,
    '60-69': 65,
    '70-79': 75,
    '80+': 80
}
members_df['AgeAtFirstClaim'] = members_df['AgeAtFirstClaim'].map(age_mapping)

# Remove rows where the 'AgeAtFirstClaim' column is missing
members_df = members_df[members_df['AgeAtFirstClaim'].notna()]


## Part d: Save the data after processing
Finally, sort the data by id and save as csv file.

# Sort the DataFrame by MemberID
members_df = members_df.sort_values(by='MemberID')

# Save the processed members' information to a CSV file
members_df.to_csv('../data/processed/processed_Member_df.csv', index=False)

# display the data after processing
new_members_df = pd.read_csv('../data/processed/processed_Member_df.csv')
new_members_df
