==> notebooks/2-claim_processing.ipynb <==
# 2. Pre-processing for Claims data
This notebook will load the claims data, then split in years, do features transformation, merge data by Id as the preprocessing.

## Part a: Data import and split by years
First, load the claims data from a CSV file.

import pandas as pd

data_path = '../data/raw/Claims.csv'
claims_df = pd.read_csv(data_path)
claims_df

# Split data by year
claims_Y1_df = claims_df[claims_df['Year'] == 'Y1']
claims_Y2_df = claims_df[claims_df['Year'] == 'Y2']
claims_Y3_df = claims_df[claims_df['Year'] == 'Y3']
claims_Y3_df

## Part b: Count Number of Claims for Each Member
We create the `ClaimsCount` column:

# Step 2: Count Number of Claims for Each Member
def calculate_claims_count(df):
    claims_count = df.groupby('MemberID').size().reset_index(name='ClaimsCount')
    return claims_count

# Calculate claims count for each year
claims_count_Y1 = calculate_claims_count(claims_Y1_df)
claims_count_Y2 = calculate_claims_count(claims_Y2_df)
claims_count_Y3 = calculate_claims_count(claims_Y3_df)
claims_count_Y1

## Part c: Count Unique Values for: `ProviderID`, `Vendor`, and `PCP`

# function for count unique values
def unique_values_count(df, feature):
    unique_count = df.groupby('MemberID')[feature].nunique().reset_index(name=f'Unique_{feature}')
    return unique_count

claims_Y1_providers = unique_values_count(claims_Y1_df, 'ProviderID')
claims_Y2_providers = unique_values_count(claims_Y2_df, 'ProviderID')
claims_Y3_providers = unique_values_count(claims_Y3_df, 'ProviderID')
claims_Y1_providers

claims_Y1_vendors = unique_values_count(claims_Y1_df, 'Vendor')
claims_Y2_vendors = unique_values_count(claims_Y2_df, 'Vendor')
claims_Y3_vendors = unique_values_count(claims_Y3_df, 'Vendor')
claims_Y1_vendors

claims_Y1_pcps = unique_values_count(claims_Y1_df, 'PCP')
claims_Y2_pcps = unique_values_count(claims_Y2_df, 'PCP')
claims_Y3_pcps = unique_values_count(claims_Y3_df, 'PCP')
claims_Y1_pcps

## Part d: Do One-Hot Encoding for Categorical Features: 
`Specialty`, `PlaceSvc`, `PrimaryConditionGroup`, `CharlsonIndex`, `ProcedureGroup`, `DSFS`

# Step 4: One-Hot Encoding for Categorical Features
def one_hot_encode(df, column):
    one_hot_encoded = pd.get_dummies(df[column])
    one_hot_encoded['MemberID'] = df['MemberID']
    one_hot_agg = one_hot_encoded.groupby('MemberID').sum().reset_index()
    return one_hot_agg

# One-hot encoding of Specialty
claims_Y1_specialty = one_hot_encode(claims_Y1_df, 'Specialty')
claims_Y2_specialty = one_hot_encode(claims_Y2_df, 'Specialty')
claims_Y3_specialty = one_hot_encode(claims_Y3_df, 'Specialty')
claims_Y1_specialty

# One-hot encoding of PlaceSvc
claims_Y1_placesvc = one_hot_encode(claims_Y1_df, 'PlaceSvc')
claims_Y2_placesvc = one_hot_encode(claims_Y2_df, 'PlaceSvc')
claims_Y3_placesvc = one_hot_encode(claims_Y3_df, 'PlaceSvc')
claims_Y1_placesvc

# One-hot encoding of PrimaryConditionGroup
claims_Y1_primary_cond = one_hot_encode(claims_Y1_df, 'PrimaryConditionGroup')
claims_Y2_primary_cond = one_hot_encode(claims_Y2_df, 'PrimaryConditionGroup')
claims_Y3_primary_cond = one_hot_encode(claims_Y3_df, 'PrimaryConditionGroup')
claims_Y1_primary_cond

# One-hot encoding of CharlsonIndex
claims_Y1_CharlsonIndex = one_hot_encode(claims_Y1_df, 'CharlsonIndex')
claims_Y2_CharlsonIndex = one_hot_encode(claims_Y2_df, 'CharlsonIndex')
claims_Y3_CharlsonIndex = one_hot_encode(claims_Y3_df, 'CharlsonIndex')
claims_Y1_CharlsonIndex

# One-hot encoding of ProcedureGroup
claims_Y1_ProcedureGroup = one_hot_encode(claims_Y1_df, 'ProcedureGroup')
claims_Y2_ProcedureGroup = one_hot_encode(claims_Y2_df, 'ProcedureGroup')
claims_Y3_ProcedureGroup = one_hot_encode(claims_Y3_df, 'ProcedureGroup')
claims_Y1_ProcedureGroup

# One-hot encoding of DSFS
claims_Y1_DSFS = one_hot_encode(claims_Y1_df, 'DSFS')
claims_Y2_DSFS = one_hot_encode(claims_Y2_df, 'DSFS')
claims_Y3_DSFS = one_hot_encode(claims_Y3_df, 'DSFS')
claims_Y1_DSFS

## Part e: Summing Numeric Feature: `PayDelay`

# Function to sum numeric feature, replace '162+' with 162, and convert to integer
def sum_numeric_feature(df, feature):
    df[feature] = df[feature].replace('162+', '162').astype(int)
    numeric_sum = df.groupby('MemberID')[feature].sum().reset_index()
    return numeric_sum

# Sum PayDelay for each member
claims_Y1_paydelay = sum_numeric_feature(claims_Y1_df, 'PayDelay')
claims_Y2_paydelay = sum_numeric_feature(claims_Y2_df, 'PayDelay')
claims_Y3_paydelay = sum_numeric_feature(claims_Y3_df, 'PayDelay')
claims_Y1_paydelay

## Part f: Merge all Features by ID and Save in Years

# Merge function, used to merge each feature by year
def merge_data_by_year(claims_count, providers, vendors, pcps, paydelay, specialty, placesvc, 
                       primary_cond, charlson_index, procedure_group, dsfs):
    from functools import reduce
    data_frames = [claims_count, providers, vendors, pcps, paydelay, specialty, placesvc,
                   primary_cond, charlson_index, procedure_group, dsfs]
    merged_data = reduce(lambda left, right: pd.merge(left, right, on='MemberID', how='outer'), data_frames)
    return merged_data

# merge data by years
merged_Y1 = merge_data_by_year(claims_count_Y1, claims_Y1_providers, claims_Y1_vendors, claims_Y1_pcps, 
                               claims_Y1_paydelay, claims_Y1_specialty, claims_Y1_placesvc, claims_Y1_primary_cond, 
                               claims_Y1_CharlsonIndex, claims_Y1_ProcedureGroup, claims_Y1_DSFS)
merged_Y2 = merge_data_by_year(claims_count_Y2, claims_Y2_providers, claims_Y2_vendors, claims_Y2_pcps, 
                               claims_Y2_paydelay, claims_Y2_specialty, claims_Y2_placesvc, claims_Y2_primary_cond, 
                               claims_Y2_CharlsonIndex, claims_Y2_ProcedureGroup, claims_Y2_DSFS)
merged_Y3 = merge_data_by_year(claims_count_Y3, claims_Y3_providers, claims_Y3_vendors, claims_Y3_pcps, 
                               claims_Y3_paydelay, claims_Y3_specialty, claims_Y3_placesvc, claims_Y3_primary_cond, 
                               claims_Y3_CharlsonIndex, claims_Y3_ProcedureGroup, claims_Y3_DSFS)

merged_Y1

# Save the merged data for each year as a CSV file
merged_Y1.to_csv('../data/processed/processed_claim_Y1.csv', index=False)
merged_Y2.to_csv('../data/processed/processed_claim_Y2.csv', index=False)
merged_Y3.to_csv('../data/processed/processed_claim_Y3.csv', index=False)

# display the data after processing
new_members_df = pd.read_csv('../data/processed/processed_claim_Y1.csv')
new_members_df
