==> notebooks/3-Drug&LabCount_processing.ipynb <==
# 3. Pre-processing for DrugCount & LabCount data
This notebook will load DrugCount & LabCount data, then split in years, do features transformation, sum data by Id as the preprocessing.

## Part a: Data import, feature type transformation, and split by years
First, load DrugCount & LabCount data from CSV files.

import pandas as pd

DrugCount_df = pd.read_csv('../data/raw//DrugCount.csv')
LabCount_df = pd.read_csv('../data/raw//LabCount.csv')

Replace the special string with an integer and convert others to integer also

# Replace string 7+ and 10+ as integer
DrugCount_df['DrugCount'].replace({'7+': '7'}, inplace=True)
LabCount_df['LabCount'].replace({'10+': '10'}, inplace=True)

# transform other numbers recorded in str as int
DrugCount_df['DrugCount'] = pd.to_numeric(DrugCount_df['DrugCount'], downcast='integer')
LabCount_df['LabCount'] = pd.to_numeric(LabCount_df['LabCount'], downcast='integer')

# Group by year
DrugCount_Y1 = DrugCount_df[DrugCount_df['Year'] == 'Y1']
DrugCount_Y2 = DrugCount_df[DrugCount_df['Year'] == 'Y2']
DrugCount_Y3 = DrugCount_df[DrugCount_df['Year'] == 'Y3']

LabCount_Y1 = LabCount_df[LabCount_df['Year'] == 'Y1']
LabCount_Y2 = LabCount_df[LabCount_df['Year'] == 'Y2']
LabCount_Y3 = LabCount_df[LabCount_df['Year'] == 'Y3']



# Display data set
DrugCount_Y1

LabCount_Y1

## Part b: Summing Numeric Feature: `DrugCount` & `LabCount`

In each year, calculate the sum of amount for two features by member id. Drop other features.

# Function to sum numeric feature
def sum_numeric_feature(df, feature):
    numeric_sum = df.groupby('MemberID')[feature].sum().reset_index()
    return numeric_sum

# Calculate drug counts and lab test counts for each member in each year 
DrugCount_summary_Y1 = sum_numeric_feature(DrugCount_Y1, 'DrugCount')
DrugCount_summary_Y2 = sum_numeric_feature(DrugCount_Y2, 'DrugCount')
DrugCount_summary_Y3 = sum_numeric_feature(DrugCount_Y3, 'DrugCount')

LabCount_summary_Y1 = sum_numeric_feature(LabCount_Y1, 'LabCount')
LabCount_summary_Y2 = sum_numeric_feature(LabCount_Y2, 'LabCount')
LabCount_summary_Y3 = sum_numeric_feature(LabCount_Y3, 'LabCount')



# Display data set
DrugCount_summary_Y1

LabCount_summary_Y1

## Part c: Merge all Features by ID and Save in Years

def merge_annual_data(drug_data, lab_data, year):
    # Use outer join to merge data
    merged_data = pd.merge(drug_data, lab_data, on='MemberID', how='outer', suffixes=('_Drug', '_Lab'))
    # If there are NaNs in the data, fill them with 0
    merged_data.fillna(0, inplace=True)
    merged_data.to_csv(f'../data/processed/processed_lab&drug_Y{year}.csv', index=False)

# Call function to merge data
merge_annual_data(DrugCount_summary_Y1, LabCount_summary_Y1, 1)
merge_annual_data(DrugCount_summary_Y2, LabCount_summary_Y2, 2)
merge_annual_data(DrugCount_summary_Y3, LabCount_summary_Y3, 3)

# Display data set
y1_df = pd.read_csv('../data/processed/processed_lab&drug_Y1.csv')
y1_df
