==> notebooks/4-merge_data.ipynb <==
# 4.Merge All Needed Data Sets

This notebook will merge all processed features data sets and label data sets for year 2 and year 3.

## Part a: Data import
First, load the data from CSV files, and then merge by member id using outer join.

import pandas as pd

# Step 1: Data Loading
# Load the members' information from the CSV file
members_df = pd.read_csv('../data/processed/processed_Member_df.csv')

claims_Y2_df = pd.read_csv('../data/processed/processed_claim_Y2.csv')
claims_Y3_df = pd.read_csv('../data/processed/processed_claim_Y3.csv')

lab_drug_Y2_df = pd.read_csv('../data/processed/processed_lab&drug_Y2.csv')
lab_drug_Y3_df = pd.read_csv('../data/processed/processed_lab&drug_Y3.csv')

days_hospital_Y2_df = pd.read_csv('../data/raw/DaysInHospital_Y2.csv')
days_hospital_Y3_df = pd.read_csv('../data/raw/DaysInHospital_Y3.csv')

# Merge data by MemberID for year 2
merged_data_Y2 = members_df.merge(claims_Y2_df, on='MemberID', how='outer').merge(lab_drug_Y2_df, on='MemberID', how='outer')
merged_data_Y2 

# Merge data by MemberID for year 3
merged_data_Y3 = members_df.merge(claims_Y3_df, on='MemberID', how='outer').merge(lab_drug_Y3_df, on='MemberID', how='outer')
merged_data_Y3

## Part b: Missing value handling
Excessive missing features are considered invalid data, and other missing values are filled with 0.

# Set condition: at least half of the features in each row need to be non-null, otherwise the row is deleted
half_features_count = len(merged_data_Y2.columns) // 2

# Delete rows where more than half of the data is empty
merged_data_Y2.dropna(thresh=half_features_count, inplace=True)
merged_data_Y3.dropna(thresh=half_features_count, inplace=True)

# Replace the NaN value with the 0 default value
merged_data_Y2.fillna(0, inplace=True)
merged_data_Y3.fillna(0, inplace=True)

# Display data for year2
merged_data_Y2

# Display data for year3
merged_data_Y3

## Part c: Insert label: `DaysInHospital` and Save Data
Merge features and label as one data set for year 2 and 3 using inner join to keep label correctness.

# Merge the year 2 data
merged_data_Y2_with_label = pd.merge(merged_data_Y2, days_hospital_Y2_df[['MemberID', 'DaysInHospital']], on='MemberID', how='inner')
merged_data_Y2_with_label

# Merge the year 3 data
merged_data_Y3_with_label = pd.merge(merged_data_Y3, days_hospital_Y3_df[['MemberID', 'DaysInHospital']], on='MemberID', how='inner')
merged_data_Y3_with_label

# Save the merged data as CSV files
merged_data_Y2_with_label.to_csv('../data/processed/merged_data_Y2.csv', index=False)
merged_data_Y3_with_label.to_csv('../data/processed/merged_data_Y3.csv', index=False)
