==> notebooks/6-rf_feature_selection.ipynb <==
# 6. Feature Selection by Random Forest Regression

The feature importance obtained through the random forest model can help us screen out the most useful features, thereby simplifying the model, reducing overfitting, and improving computational efficiency.

## Part a: Data import
First, load the data from CSV files, and get data as features and label.

import pandas as pd
import numpy as np
df = pd.read_csv('../data/processed/merged_data_Y2.csv')
df_y3 = pd.read_csv('../data/processed/merged_data_Y3.csv')
df

# Remove columns not used for prediction
X = df.drop(columns=['MemberID', 'DaysInHospital'])

# Use DaysInHospital as a label
y = df['DaysInHospital']
X

## Part b: Training and evaluating the model before feature selection

from sklearn.ensemble import RandomForestClassifier
from sklearn.feature_selection import SelectFromModel
from sklearn.model_selection import train_test_split

# Divide into training set and test set
X_train, X_test, y_train, y_test = train_test_split(X, y, test_size=0.2, random_state=42)

# Train the random forest model
rf = RandomForestClassifier(n_estimators=100, random_state=42)
rf.fit(X_train, y_train)

# Evaluate model performance before feature selection
pre_selection_score = rf.score(X_test, y_test)
print(f"Coefficient of Determination (R²) before feature selection: {pre_selection_score}")

## Part c: Feature selection based on feature importance

# Select features whose feature importance is higher than the average
selector = SelectFromModel(rf, threshold="mean")
selector.fit(X_train, y_train)

# Get selected features
selected_features = selector.get_support(indices=True)
selected_feature_names = X.columns[selected_features]

# Print the selected feature names
print("Selected features:")
print(selected_feature_names)

# Extract selected features
X_train_selected = selector.transform(X_train)
X_test_selected = selector.transform(X_test)

## Part d: Model after feature selection

# Retrain the model using the selected features
rf_selected = RandomForestClassifier(n_estimators=100, random_state=42)
rf_selected.fit(X_train_selected, y_train)

# Evaluate model performance after feature selection
post_selection_score = rf_selected.score(X_test_selected, y_test)
print(f"Coefficient of Determination (R²) after feature selection: {post_selection_score}")

An increase in R² after removing features usually means that unnecessary, irrelevant, or negatively impacting features have been removed. In this way, the model complexity is reduced, overfitting is avoided, and the generalization ability of the model is improved, making it perform better on the test set. 

## Part e: Save data after feature selection

# apply feature selection on the year 2 data
rf_data_y2 = df[['AgeAtFirstClaim', 'ClaimsCount', 'Unique_ProviderID', 'Unique_Vendor',
       'PayDelay', 'Diagnostic Imaging', 'Emergency', 'General Practice',
       'Internal', 'Surgery', 'Ambulance', 'Independent Lab',
       'Inpatient Hospital', 'Office', 'Urgent Care', 'MSC2a3', 'ROAMI',
       'TRAUMA', '1-2', 'EM', 'MED', 'PL', 'RAD', 'SDS', '0- 1 month',
       'DrugCount', 'LabCount', 'DaysInHospital']]

# store the data
rf_data_y2.to_csv('../data/train_test/rf_data_Y2.csv', index=False)
rf_data_y2

# apply feature selection on the year 3 data
rf_data_y3 = df_y3[['AgeAtFirstClaim', 'ClaimsCount', 'Unique_ProviderID', 'Unique_Vendor',
       'PayDelay', 'Diagnostic Imaging', 'Emergency', 'General Practice',
       'Internal', 'Surgery', 'Ambulance', 'Independent Lab',
       'Inpatient Hospital', 'Office', 'Urgent Care', 'MSC2a3', 'ROAMI',
       'TRAUMA', '1-2', 'EM', 'MED', 'PL', 'RAD', 'SDS', '0- 1 month',
       'DrugCount', 'LabCount', 'DaysInHospital']]

# store the data
rf_data_y3.to_csv('../data/train_test/rf_data_Y3.csv', index=False)
rf_data_y3
