==> notebooks/11-linear_regression.ipynb <==
import pandas as pd
from sklearn.model_selection import train_test_split
from sklearn.linear_model import LinearRegression
from sklearn.preprocessing import StandardScaler, OneHotEncoder
from sklearn.compose import ColumnTransformer
from sklearn.pipeline import Pipeline
from sklearn.metrics import mean_squared_error, mean_absolute_error, r2_score
from sklearn.linear_model import Ridge
from sklearn.model_selection import GridSearchCV, KFold
import numpy as np

dataY2 = pd.read_csv('../data/processed/merged_data_Y2.csv')
dataY3 = pd.read_csv('../data/processed/merged_data_Y3.csv')

#model_1,feacture1 in linear regression
#model_2,feacture2 in linear regression
#model_3,feacture3 in linear regression

#model_4,feacture1 in linear regression with Ridge
#model_5,feacture2 in linear regression with Ridge
#model_6,feacture3 in linear regression with Ridge

#model_7,feature1, convert all negative to 0
#model_8,feature2, convert all negative to 0
#model_9,feature3, convert all negative to 0

#model_1,feacture1 in linear regression

selected_features_1 = ['ClaimsCount', 'PayDelay', 'Unique_ProviderID', 'MED', 'EM', 'SCS', 'RAD', 'ANES', 'SDS', 
                     'PL', 'SIS', 'SMS', 'SGS', 'SEOA', 'SNS', 'SAS', 'SRS', 'SUS', 'SO', 'SMCD', 'Office', 
                     'Independent Lab', 'Outpatient Hospital', 'Inpatient Hospital', 'Urgent Care', 'Ambulance', 
                     'Home', 'Other_y', 'Surgery', 'Internal', 'Laboratory', 'Pediatrics', 'Rehabilitation', 
                     'Diagnostic Imaging', 'Anesthesiology', 'Emergency', 'General Practice', 'Other_x', 
                     'Obstetrics and Gynecology', 'Pathology']

X_train = dataY2[selected_features_1]
y_train = dataY2['DaysInHospital']

X_test = dataY3[selected_features_1]
y_test = dataY3['DaysInHospital']

scaler = StandardScaler()

X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)

model = LinearRegression()
model.fit(X_train_scaled, y_train)

y_pred = model.predict(X_test_scaled)

mse = mean_squared_error(y_test, y_pred)
r2 = r2_score(y_test, y_pred)

print(f"Mean Squared Error (MSE) on Y3: {mse}")
print(f"R² Score on Y3: {r2}")

results = pd.DataFrame({
    'Actual DaysInHospital': y_test,
    'Predicted DaysInHospital': y_pred
})
print(results.head())

#model_2,feacture2 in linear regression

selected_features_2 = ['ClaimsCount', 'Diagnostic Imaging', 'Emergency', 'Pathology',
       'Ambulance', 'Home', 'Inpatient Hospital', 'Office', 'Other_y',
       'Outpatient Hospital', 'Urgent Care', 'APPCHOL', 'CANCRM', 'CATAST',
       'LIVERDZ', 'PERINTL', 'PERVALV', 'PNCRDZ', 'RENAL1', 'SEPSIS', 'STROKE',
       '0', '1-2', '3-4', '5+', 'ANES', 'PL', 'SAS', 'SCS', 'SDS', 'SMCD',
       'SMS', 'SO']

X_train = dataY2[selected_features_2]
y_train = dataY2['DaysInHospital']

X_test = dataY3[selected_features_2]
y_test = dataY3['DaysInHospital']

scaler = StandardScaler()

X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)

model = LinearRegression()
model.fit(X_train_scaled, y_train)

y_pred = model.predict(X_test_scaled)

mse = mean_squared_error(y_test, y_pred)
mae = mean_absolute_error(y_test, y_pred)
rmse = np.sqrt(mse)
r2 = r2_score(y_test, y_pred)

print(f"Mean Squared Error (MSE) on Y3: {mse}")
print(f"R² Score on Y3: {r2}")

results = pd.DataFrame({
    'Actual DaysInHospital': y_test,
    'Predicted DaysInHospital': y_pred
})
print(results.head())

#model_3,feacture3 in linear regression

selected_features_3 = ['AgeAtFirstClaim', 'ClaimsCount', 'Unique_ProviderID', 'Unique_Vendor', 'PayDelay',
    'Diagnostic Imaging', 'Emergency', 'General Practice', 'Internal', 'Surgery',
    'Ambulance', 'Independent Lab', 'Inpatient Hospital', 'Office', 'Urgent Care',
    'MSC2a3', 'ROAMI', 'TRAUMA', '1-2', 'EM', 'MED', 'PL', 'RAD', 'SDS',
    '0- 1 month', 'DrugCount', 'LabCount']

X_train = dataY2[selected_features_3]
y_train = dataY2['DaysInHospital']

X_test = dataY3[selected_features_3]
y_test = dataY3['DaysInHospital']

scaler = StandardScaler()

X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)

model = LinearRegression()
model.fit(X_train_scaled, y_train)

y_pred = model.predict(X_test_scaled)

mse = mean_squared_error(y_test, y_pred)
mae = mean_absolute_error(y_test, y_pred)
rmse = np.sqrt(mse)
r2 = r2_score(y_test, y_pred)

print(f"Mean Squared Error (MSE) on Y3: {mse}")
print(f"R² Score on Y3: {r2}")

results = pd.DataFrame({
    'Actual DaysInHospital': y_test,
    'Predicted DaysInHospital': y_pred
})
print(results.head())

#model_4,feacture1 in linear regression with Ridge

X_train = dataY2[selected_features_1]
y_train = dataY2['DaysInHospital']

X_test = dataY3[selected_features_1]
y_test = dataY3['DaysInHospital']

scaler = StandardScaler()

X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)

ridge_model = Ridge()

alpha_values = np.logspace(-3, 3, 50) 

param_grid = {'alpha': alpha_values}
kfold = KFold(n_splits=5, shuffle=True, random_state=10)
grid_search = GridSearchCV(estimator=ridge_model, param_grid=param_grid, cv=kfold, scoring='neg_mean_squared_error')

grid_search.fit(X_train_scaled, y_train)

best_alpha = grid_search.best_params_['alpha']
best_ridge_model = grid_search.best_estimator_

print(f"Best alpha found through GridSearchCV: {best_alpha}")

y_pred = best_ridge_model.predict(X_test_scaled)


mse = mean_squared_error(y_test, y_pred)
r2 = r2_score(y_test, y_pred)

print(f"Mean Squared Error (MSE) on Y3 with Ridge (Best Alpha {best_alpha}): {mse}")
print(f"R² Score on Y3 with Ridge (Best Alpha {best_alpha}): {r2}")

results = pd.DataFrame({
    'Actual DaysInHospital': y_test,
    'Predicted DaysInHospital': y_pred
})
print(results.head())

#model_5,feacture2 in linear regression with Ridge

X_train = dataY2[selected_features_2]
y_train = dataY2['DaysInHospital']

X_test = dataY3[selected_features_2]
y_test = dataY3['DaysInHospital']

scaler = StandardScaler()

X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)

ridge_model = Ridge()

alpha_values = np.logspace(-3, 3, 50) 

param_grid = {'alpha': alpha_values}
kfold = KFold(n_splits=5, shuffle=True, random_state=10)
grid_search = GridSearchCV(estimator=ridge_model, param_grid=param_grid, cv=kfold, scoring='neg_mean_squared_error')

grid_search.fit(X_train_scaled, y_train)

best_alpha = grid_search.best_params_['alpha']
best_ridge_model = grid_search.best_estimator_

print(f"Best alpha found through GridSearchCV: {best_alpha}")

y_pred = best_ridge_model.predict(X_test_scaled)


mse = mean_squared_error(y_test, y_pred)
r2 = r2_score(y_test, y_pred)

print(f"Mean Squared Error (MSE) on Y3 with Ridge (Best Alpha {best_alpha}): {mse}")
print(f"R² Score on Y3 with Ridge (Best Alpha {best_alpha}): {r2}")

results = pd.DataFrame({
    'Actual DaysInHospital': y_test,
    'Predicted DaysInHospital': y_pred
})
print(results.head())

#model_6,feacture3 in linear regression with Ridge

X_train = dataY2[selected_features_3]
y_train = dataY2['DaysInHospital']

X_test = dataY3[selected_features_3]
y_test = dataY3['DaysInHospital']

scaler = StandardScaler()

X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)

ridge_model = Ridge()

alpha_values = np.logspace(-3, 3, 50) 

param_grid = {'alpha': alpha_values}
kfold = KFold(n_splits=5, shuffle=True, random_state=10)
grid_search = GridSearchCV(estimator=ridge_model, param_grid=param_grid, cv=kfold, scoring='neg_mean_squared_error')

grid_search.fit(X_train_scaled, y_train)

best_alpha = grid_search.best_params_['alpha']
best_ridge_model = grid_search.best_estimator_

print(f"Best alpha found through GridSearchCV: {best_alpha}")

y_pred = best_ridge_model.predict(X_test_scaled)


mse = mean_squared_error(y_test, y_pred)
r2 = r2_score(y_test, y_pred)

print(f"Mean Squared Error (MSE) on Y3 with Ridge (Best Alpha {best_alpha}): {mse}")
print(f"R² Score on Y3 with Ridge (Best Alpha {best_alpha}): {r2}")

results = pd.DataFrame({
    'Actual DaysInHospital': y_test,
    'Predicted DaysInHospital': y_pred
})
print(results.head())

#model_7,feature1, convert all negative to 0

X_train = dataY2[selected_features_1]
y_train = dataY2['DaysInHospital']

X_test = dataY3[selected_features_1]
y_test = dataY3['DaysInHospital']

scaler = StandardScaler()

X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)

model = LinearRegression()
model.fit(X_train_scaled, y_train)

y_pred = model.predict(X_test_scaled)

y_pred = np.maximum(0, y_pred)  

mse = mean_squared_error(y_test, y_pred)
r2 = r2_score(y_test, y_pred)

print(f"Mean Squared Error (MSE) on Y3: {mse}")
print(f"R² Score on Y3: {r2}")

results = pd.DataFrame({
    'Actual DaysInHospital': y_test,
    'Predicted DaysInHospital': y_pred
})
print(results.head())

#model_8,feature2, convert all negative to 0

X_train = dataY2[selected_features_2]
y_train = dataY2['DaysInHospital']

X_test = dataY3[selected_features_2]
y_test = dataY3['DaysInHospital']

scaler = StandardScaler()

X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)

model = LinearRegression()
model.fit(X_train_scaled, y_train)

y_pred = model.predict(X_test_scaled)

y_pred = np.maximum(0, y_pred)  

mse = mean_squared_error(y_test, y_pred)
r2 = r2_score(y_test, y_pred)

print(f"Mean Squared Error (MSE) on Y3: {mse}")
print(f"R² Score on Y3: {r2}")

results = pd.DataFrame({
    'Actual DaysInHospital': y_test,
    'Predicted DaysInHospital': y_pred
})
print(results.head())

#model_9,feature3, convert all negative to 0

X_train = dataY2[selected_features_3]
y_train = dataY2['DaysInHospital']

X_test = dataY3[selected_features_3]
y_test = dataY3['DaysInHospital']

scaler = StandardScaler()

X_train_scaled = scaler.fit_transform(X_train)
X_test_scaled = scaler.transform(X_test)

model = LinearRegression()
model.fit(X_train_scaled, y_train)

y_pred = model.predict(X_test_scaled)

y_pred = np.maximum(0, y_pred)  

mse = mean_squared_error(y_test, y_pred)
r2 = r2_score(y_test, y_pred)

print(f"Mean Squared Error (MSE) on Y3: {mse}")
print(f"R² Score on Y3: {r2}")

results = pd.DataFrame({
    'Actual DaysInHospital': y_test,
    'Predicted DaysInHospital': y_pred
})
print(results.head())
